fix cast

+++ src/mmpars.c
@@ -163,7 +163,7 @@ char *readRawSource(vstring fileBuf, long *size) {
   if (fbPtr != fileBuf + charCount) {
     /* To help debugging: */
     printf("fbPtr=%p fileBuf=%p charCount=%ld diff=%ld\n",
-        (void*)fbPtr, (void*)fileBuf, charCount, fbPtr - fileBuf - charCount);
+        (void*)fbPtr, (void*)fileBuf, charCount, (long)(fbPtr - fileBuf) - charCount);
     bug(1704);
   }
